requirements.txt
fastapi
uvicorn
easyocr
pytesseract
pillow
torch
transformers
spacy
numpy
pdf2image
pdfminer.six
opencv-python